SKSE related update.
Mod now resides in Data/Plugins/Sumwunn which is then loaded by a dll called by SKSE residing in SKSE/Plugins.

This is done for the following reasons:

This allows me to host just one file on the Nexus preventing confusion.

If someone has both the dll loader and the SKSE installed two instances of AchievementsModsEnabler could be loaded.
That would lead to the log file saying NO because a second instance was loaded and couldn't find the bytes to patch because they were already patched.